Share CIDR routing to container method
