+++ newsfragments/1220.bugfix
@@ -1,0 +1,2 @@
+Setup sshuttle in Container method to only route cluster CIDR like the VPN
+method.
